+++ src/components/Nav.js
@@ -1,5 +1,5 @@
 import React from 'react';
-// import { Link } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 const Nav = () => {
   if(localStorage.username){
