+++ src/containers/App/App.css
@@ -1,0 +1,12 @@
+.App {
+  background-color: pink;
+}
+
+.App-header {
+  border: 1px solid black;
+}
+
+.App-title {
+  border: 1px solid black;
+}
+
