+++ src/index.js
@@ -1,13 +1,19 @@
+/*REACT ENTRY POINT*/
+/*REACT ENTRY POINT*/
+/*REACT ENTRY POINT*/
+
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware } from 'redux';
-import App from './containers/App';
-import reducers from './reducers';
 import { connect } from 'react-redux';
+
+import reducers from './reducers';
 import thunk from 'redux-thunk';
 import registerServiceWorker from './lib/registerServiceWorker';
 
+/*COMPONENTS*/
+import App from './containers/App/App';
 
 // --STORE----------///////
 const store = createStore(
